Create Basic Stack Implementation using Class & Object Concept
I have created a class named stack in which there are 4 methods __init__ for creating the structure of the stack, then there are push, pop, peek which are used for inserting, removing and displaying the last inserted value in the stack. I have also created an additional function named operation() which gives a menu of the operations that the user can perform on the stack.